pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "atlas-api"
version = "0.1.0"
description = "3D mapping and reconstruction API"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "fastapi==0.109.0",
    "uvicorn[standard]==0.27.0",
    "python-multipart==0.0.6",
    "pydantic-settings==2.1.0",
    "celery==5.3.4",
    "redis==5.0.1",
    "Pillow==10.2.0",
    "opencv-python==4.9.0.80",
    "numpy==1.26.3",
    "open3d==0.19.0",
    "pydantic==2.5.3",
    "python-jose[cryptography]==3.3.0",
    "trimesh>=4.6.13",
    # SAM + ML stack (CPU builds by default)
    "torch>=2.2.0,<3.0.0",
    "torchvision>=0.17.0,<1.0.0",
    "segment-anything @ git+https://github.com/facebookresearch/segment-anything.git",
    # COLMAP via Python wrapper
    "pycolmap>=0.3.1",
]

[project.optional-dependencies]
dev = [
    "pytest==7.4.4",
    "pytest-asyncio==0.23.3",
]

[tool.setuptools]
packages = ["api", "worker"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
